Optimizing Movement - Part 3
Got the movement dot to stay within it's boundaries.  Additionally, an issue involving the dot from resetting back to the appropriate anchor when an axis wasn't active was found AND solved.  Next up is to see whether or not the speed is doubling when two opposing axis activations occur (WA, WD, SA, SD).