Optimizing Movement - Part 4
Decided to leave the illusion testing for another day because the more and more I see the illusion the more and more I believe it truly is JUST an illusion.  Instead, I worked on applying the properties of mass, push and pull to the Player Object.  In the end, it caused the Player Object's movement to become highly varied based on it's properties.  Me happy. :)