Optimizing Movement - Part 5
Added the property "turn" to the player object.  Allows one to control the time it takes to align to a player instructed direction (i.e. WASD).  Additionally, I confirmed that the previous stated illusion was in fact just an illusion.  Lastly, the plan for next session is to look into some minor optimizations and finalize the naming conventions for properties within the worked objects.