Cannon Track-Ring Development - Part 2
Got the weapon object to stick on the ring's outer edge even if the cursor is within the ring's sprite.  Additionally, it follows the player now.  So it looks like a fancy fucking object that I'm proud of.  Next up is to have the ring rotate left or right based on the Q and E key.  Very pumped right now.  Me do good. :)