Cannon Track-Ring Development - Part 5
Created speed property for the ring and decided to go with the ring moving instead of the weapon moving on the ring.  The reason for the decision was because my brain started hurting trying to solve problems related to 360 degree movement.  It's not worth the effort for the starting cannon.  One day I'll come back to that issue, one day... Until then we'll have a ring that moves by a set speed through holding down space. :)