Cannon Track-Ring Development - Part 6
Re-hauled the original Cannon Diagram of only having a ring, weapon and the cannon chassis itself.  Now there is a Ring, Track, Chassis, Ring Anchor, Track Anchor and weapons are called Attachments.  I did this because the original concept incorporated a weapon track (the thing the weapon is connected to or rests on) that was integrated into the ring.  In the end, this proved to be problematic and made the code very difficult to both design and program.  However, incorporating anchors and an official track object gives a far more easier approach to designing cannons since they're now split.  If anyone is worried, this does nothing to harm the original cannon concept functionality wise and has the option to visually duplicate the original too.  All in all, a lot changed for the better.