Crafting the 1st Weapon - Part 2
The weapon angle dictates the bullet direction and projectiles now move based on their range, which is augmented by the attack_item (a weapon) they are connected to.  Projectiles act independently from a weapon.  A weapon merely changes a projectile's stats using it's own built-in stats.

Next thing to work on is further defining what the stats will do on the weapon. :)