Crafting the 1st Weapon - Part 3
Fire Rate and Projectile Speed stats now have functionality.  Working on damage and damage modifier stats (critical and weakspot) next, but in order to do that I have to figure out why collision isn't working inside the game.  Maybe something to do with layers?  Also multishot and recoil are on the to-do list.