Crafting the 1st Weapon - Part 4
Kick (currently referred to as knockback, but plan to change in the next session) is nearly done.  Kick currently doesn't effect the target's direction well and isn't varied based on the entity mass and force, so it's got some stuff left to do inside the stat.  Besides that Projectile Acceleration and Ammo Consumption has been established and is functioning properly.