Crafting the 1st Weapon - Part 5
Collision issue has been solved.  Which in turned allowed for damage and critical damage to be implemented.

Also decided in this commit that weakspot damage should be calculated at a later date.  I got so far  as to even fully implement weakspot damage, but it was so messy on a functional front that I chose to scrap it.    Either way, it's not a crucial element to add at the moment (would rather not spend a century on optimizing it during this stage of development).

Lastly, the Kick angle issue was fixed to make sure the kick angle stores the angle where a shot occurs, not where the attachment is currently at.