Optimizing, Weight and Time - Part 2
Changed the naming convention of all variables to be more universal, reduce character count and increase visual clarity.  In doing so, I believe the code readability has lowered.  But I believe this route is more beneficial in the long-term and the readability was only "slightly" reduced.

Additionally, the Gun Code was slightly optimized.