Optimizing, Weight and Time - Part 3
There can now be an infinite number of attachments and slots on the Cannon.  Also I removed some unnecessary lines regarding object positioning that I was using for testing purposes in an earlier stage of development.

Next up, my goal will be to work on having all the Cannon Attachments move with the Cannon.  But only have the active one move with the cursor on the track as it's following the Cannon's position.  In doing so, I will have also created attachment weight seeing as how the property is already set and I've tested it's functionality already too.