Optimizing, Weight and Time - Part 4
Added weight to attachments.  Weight effects the speed the attachment moves on the track.  However, the implementation could definitely be optimized.  I feel that the development of it was poor, but developing it was challenging so for right now I'm happy that it's functional for now.  Will come back to it before I push the branch. :)