Optimizing, Weight and Time - Final
I was trying to redo the attachment object tree in order to accommodate for future attachments to be made, but in the process of hours of thought on the subject... I realized something.

Attachments are meant to be very unique.  Every time I thought of an object layout I figured was the "one", 5 minutes later I kept coming up with an exception to the rule.  Simply put, each Attachment will be linked to the Attachment Object and nothing more.  Instead, we'll just have to make sure that the attachments share the correct variable naming.

Now, I'd imagine one day we can objectify them.  But for right now I get the feeling it could ruin their direction if we start homogenizing from the start, so we're going to keep them unique until we can't take it anymore.