Crafting the 1st Weapon - Part 8
Falloff Point and Severity have been added.  However, severity is acting strangely when working with a negative value.  I believe I know the fix, but I don't have the time tonight.

Additionally, I have decided to ice punchthrough/phasethrough for now.  It'll take a bunch of time to learn to do and it's not exactly a REQUIRED stat to implement, but one day I would definitely love to add it into the weapon(s).

Moreover, some work needs to be done to the accuracy lines to accommodate for the width of the projectiles.  Tomorrow will just be an overall optimization day for the 1st weapon.