Crafting the 1st Weapon - Part 9
Optimized Accuracy Lines to be more visually accurate to the projectile width (specifically the width because it's the only thing that effects a barrel from a 2d perspective).  Additionally, I added in the option to change the trigger type of the gun.  Trigger Type being whether you shoot the gun by holding down the trigger, pulling the trigger once to fire a singular shot or releasing the trigger to fire a projectile.

Next up on the list is to solve all the crazy riddles involved with optimizing the gun.  Could take awhile, but I'm honestly unsure.