Cleanup
Getting things setup for the first prototype build.  Got rid of anything that's useless for the new entity code and any visuals that don't relate to the functionality of the first form of the player.  In addition, I added a function that checks to see if a specified mouse event is being triggered for the weapon blueprint.